requirements.txt
PyMuPDF==1.23.14
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
# Module 2: Semantic Indexing
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3